MBS-12648 Fix review notes
