+++ settings.gradle.kts
@@ -103,7 +103,6 @@ include(":subprojects:test-runner:test-report")
 include(":subprojects:test-runner:test-inhouse-runner")
 include(":subprojects:test-runner:test-instrumentation-runner")
 include(":subprojects:test-runner:test-robolectric-inhouse-runner")
-include(":subprojects:test-runner:test-robolectric-runner")
 include(":subprojects:test-runner:shared:logger-providers")
 include(":subprojects:test-runner:k8s-deployments-cleaner")
 include(":subprojects:test-runner:instrumentation-changed-tests-finder")
