+++ subprojects/android-test/ui-testing-core-app/src/test/kotlin/com/avito/android/test/app/core/AppRobolectricTestRunner.kt
@@ -2,4 +2,4 @@ package com.avito.android.test.app.core
 
 import com.avito.robolectric.runner.InHouseRobolectricTestRunner
 
-open class RoboTestRunner(testClass: Class<*>) : InHouseRobolectricTestRunner(testClass)
+open class AppRobolectricTestRunner(testClass: Class<*>) : InHouseRobolectricTestRunner(testClass)
