+++ subprojects/test-runner/test-robolectric-inhouse-runner/build.gradle.kts
@@ -4,7 +4,8 @@ plugins {
 }
 
 dependencies {
-    api(projects.subprojects.testRunner.testRobolectricRunner)
+    api(libs.robolectric)
+    api(libs.junit)
 
     implementation(projects.subprojects.logger.logger)
     implementation(projects.subprojects.testRunner.shared.loggerProviders)
