+++ subprojects/test-runner/test-robolectric-inhouse-runner/src/main/kotlin/com/avito/robolectric/runner/InHouseRobolectricTestRunner.kt
@@ -5,15 +5,23 @@ import com.avito.logger.LogLevel
 import com.avito.logger.LoggerFactory
 import com.avito.logger.LoggerFactoryBuilder
 import com.avito.logger.handler.PrintlnLoggingHandlerProvider
+import org.robolectric.RobolectricTestRunner
+import org.robolectric.util.inject.Injector
 
-public open class InHouseRobolectricTestRunner(
-    private val testClass: Class<*>,
-) : BaseRobolectricTestRunner(testClass) {
+public open class InHouseRobolectricTestRunner
+@JvmOverloads constructor(
+    testClass: Class<*>,
+    injector: Injector = DEFAULT_INJECTOR
+) : RobolectricTestRunner(testClass, injector) {
 
-    override val loggerFactory: LoggerFactory by lazy {
+    protected val loggerFactory: LoggerFactory by lazy {
         LoggerFactoryBuilder()
             .metadataProvider(AndroidTestLoggerMetadataProvider(testClass.name))
             .addLoggingHandlerProvider(PrintlnLoggingHandlerProvider(LogLevel.DEBUG, printStackTrace = false))
             .build()
+    }
+
+    private companion object {
+        private val DEFAULT_INJECTOR = defaultInjector().build()
     }
 }
