--- subprojects/test-runner/test-robolectric-runner/build.gradle.kts
@@ -1,10 +1,0 @@
-plugins {
-    id("convention.kotlin-jvm")
-    id("convention.publish-kotlin-library")
-}
-
-dependencies {
-    api(libs.robolectric)
-    api(libs.junit)
-    implementation(projects.subprojects.logger.logger)
-}
